pyproject.toml
[project]
name = "qwen-sft"
version = "0.1.0"
description = "Qwen SFT training project"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "datasets>=2.14.0",
    "tokenizers>=0.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "huggingface-hub>=0.16.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.23.0",
    "safetensors>=0.3.0",
    "bitsandbytes>=0.41.0",
    "flash-attn>=2.3.0",
    "xformers>=0.0.22",
    "triton>=2.0.0",
    "ninja>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 